// ambientika-local-control/src/models/enum/app-events.enum.ts
export enum AppEvents {
    LOCAL_SOCKET_CONNECTED = 'LOCAL_SOCKET_CONNECTED',
    LOCAL_SOCKET_DISCONNECTED = 'LOCAL_SOCKET_DISCONNECTED',
    LOCAL_SOCKET_DATA_UPDATE_RECEIVED = 'LOCAL_SOCKET_DATA_UPDATE_RECEIVED',
    REMOTE_SOCKET_CONNECTED = 'REMOTE_SOCKET_CONNECTED',
    REMOTE_SOCKET_DISCONNECTED = 'REMOTE_SOCKET_DISCONNECTED',
    REMOTE_SOCKET_DATA_UPDATE_RECEIVED = 'REMOTE_SOCKET_DATA_UPDATE_RECEIVED',
    DEVICE_STATUS_UPDATE_RECEIVED = 'DEVICE_STATUS_UPDATE_RECEIVED',
    DEVICE_COMMAND_UPDATE_RECEIVED = 'DEVICE_COMMAND_UPDATE_RECEIVED',
    DEVICE_OPERATING_MODE_UPDATE = 'DEVICE_OPERATING_MODE_UPDATE',
    LOCAL_SOCKET_DATA_UPDATE = 'LOCAL_SOCKET_DATA_UPDATE',
    DEVICE_ONLINE = 'DEVICE_ONLINE',
    DEVICE_OFFLINE = 'DEVICE_OFFLINE',
    DEVICE_FILTER_RESET = 'DEVICE_FILTER_RESET',
    DEVICE_WEATHER_UPDATE = 'DEVICE_WEATHER_UPDATE',
    DEVICE_BROADCAST_STATUS_RECEIVED = 'DEVICE_BROADCAST_STATUS_RECEIVED'
}